ltp file separated
