+++ src/transformers/models/__init__.py
@@ -47,6 +47,7 @@ from . import (
     led,
     longformer,
     lxmert,
+    ltp,
     m2m_100,
     marian,
     mbart,
