+++ src/transformers/models/ltp/__init__.py
@@ -22,34 +22,31 @@ from ...file_utils import _BaseLazyModule, is_torch_available
 
 
 _import_structure = {
-    "configuration_ibert": ["IBERT_PRETRAINED_CONFIG_ARCHIVE_MAP", "IBertConfig"],
+    "configuration_ltp": ["LTP_PRETRAINED_CONFIG_ARCHIVE_MAP", "LTPConfig"],
 }
 
 if is_torch_available():
-    _import_structure["modeling_ibert"] = [
-        "IBERT_PRETRAINED_MODEL_ARCHIVE_LIST",
-        "IBertForMaskedLM",
-        "IBertForMultipleChoice",
-        "IBertForQuestionAnswering",
-        "IBertForSequenceClassification",
-        "IBertForTokenClassification",
-        "IBertModel",
-        "IBertPreTrainedModel",
+    _import_structure["modeling_ltp"] = [
+        "LTP_PRETRAINED_MODEL_ARCHIVE_LIST",
+        "LTPForMaskedLM",
+        "LTPForMultipleChoice",
+        "LTPForQuestionAnswering",
+        "LTPForSequenceClassification",
+        "LTPForTokenClassification",
+        "LTPModel",
     ]
 if TYPE_CHECKING:
-    from .configuration_ibert import IBERT_PRETRAINED_CONFIG_ARCHIVE_MAP, IBertConfig
     from .configuration_ltp import LTP_PRETRAINED_CONFIG_ARCHIVE_MAP, LTPConfig
 
     if is_torch_available():
-        from .modeling_ibert import (
-            IBERT_PRETRAINED_MODEL_ARCHIVE_LIST,
-            IBertForMaskedLM,
-            IBertForMultipleChoice,
-            IBertForQuestionAnswering,
-            IBertForSequenceClassification,
-            IBertForTokenClassification,
-            IBertModel,
-            IBertPreTrainedModel,
+        from .modeling_ltp import (
+            LTP_PRETRAINED_MODEL_ARCHIVE_LIST,
+            LTPForMaskedLM,
+            LTPForMultipleChoice,
+            LTPForQuestionAnswering,
+            LTPForSequenceClassification,
+            LTPForTokenClassification,
+            LTPModel,
         )
 else:
     import importlib
